services/processor_worker/requirements.txt
aiokafka==0.13.0
asyncpg==0.31.0
prometheus-client==0.24.1
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
structlog==25.5.0
pydantic==2.12.5
pydantic-settings==2.13.1
aiohttp==3.13.3
